KaspaServices/requirements.txt
astroid==1.6.5
backports.functools-lru-cache==1.5
cassandra-driver==3.14.0
click==6.7
colorama==0.3.9
configparser==3.5.0
enum34==1.1.6
Flask==1.0.2
Flask-HTTPAuth==3.2.4
futures==3.2.0
gunicorn==19.9.0
isort==4.3.4
itsdangerous==0.24
Jinja2==2.10
lazy-object-proxy==1.3.1
MarkupSafe==1.0
mccabe==0.6.1
passlib==1.7.1
pkg-resources==0.0.0
pylint==1.9.2
python-dotenv==0.9.1
singledispatch==3.4.0.3
six==1.11.0
waitress==1.1.0
Werkzeug==0.14.1
wrapt==1.10.11
